Preserve input dtype in neuron modules and surrogate gradients
Fixes #421, fixes #422.

All surrogate gradient forward passes hard-cast spikes with
(input_ > 0).float(), downcasting float64 workflows to float32. In
recurrent neurons (RLeaky, RSynaptic) the float32 spike is fed back
into the float64 recurrent layer on the next step, raising
'mat1 and mat2 must have the same dtype'.

- surrogate.py: emit spikes as (input_ > 0).to(input_.dtype) in all
  surrogate forward passes
- neurons.py, leakyparallel.py, leakykernel.py, leakyunroll.py: same
  fix in _surrogate_bypass and the module-private ATan surrogates
- slstm.py, sconv2dlstm.py: initialize hidden states with the input
  dtype instead of the default dtype

Add tests parametrized over dtype x module and dtype x surrogate,
including the two-step recurrent case that previously crashed.

+++ snntorch/_neurons/sconv2dlstm.py
@@ -303,10 +303,14 @@ class SConv2dLSTM(SpikingNeuron):
         size = input_.size()
         correct_shape = (size[0], self.out_channels, size[2], size[3])
         if not self.syn.shape == correct_shape:
-            self.syn = torch.zeros(correct_shape, device=self.syn.device)
+            self.syn = torch.zeros(
+                correct_shape, device=self.syn.device, dtype=input_.dtype
+            )
 
         if not self.mem.shape == correct_shape:
-            self.mem = torch.zeros(correct_shape, device=self.mem.device)
+            self.mem = torch.zeros(
+                correct_shape, device=self.mem.device, dtype=input_.dtype
+            )
 
         self.reset = self.mem_reset(self.mem)
         self.syn, self.mem = self.state_function(input_)
